+++ mvn-golang-examples/mvn-golang-example-maven-repository/mvn-go-test-lib/src/com/igormaznitsa/mvngotest/SomeUtils.go
@@ -1,10 +1,13 @@
 package mvngotest
 
-import "fmt"
+import (
+	meta "com/igormaznitsa/mvngotestmeta"
+	"fmt"
+)
 
 var Buildstamp string
 var svnRevision string
 
 func SomeTestMethod() {
-	fmt.Printf("Hello from mvn-go-test-lib module!\n")
+	fmt.Printf(meta.SomeMetaMethod("mvn-go-test-lib module") + "!\n")
 }