+++ mvn-golang-examples/mvn-golang-example-maven-repository/mvn-go-test-metalib/src/com/igormaznitsa/mvngotestmeta/MetaUtils.go
@@ -1,0 +1,5 @@
+package mvngotestmeta
+
+func SomeMetaMethod(arg string) string {
+	return "Hello " + arg
+}